+++ src/main/kotlin/com/github/copyinaction/performance/repository/PerformanceRepository.kt
@@ -12,4 +12,6 @@ interface PerformanceRepository : JpaRepository<Performance, Long> {
 
     @EntityGraph(attributePaths = ["venue", "company"])
     override fun findAll(): List<Performance>
+
+    fun existsByVenueId(venueId: Long): Boolean
 }